Move url helpers to util.js
