+++ src/components/project/recent.js
@@ -9,7 +9,7 @@ import { ProjectFileList } from './file.js'
 import { consolidate, reload } from '../../slices/project-files.js'
 import * as act from '../../actions/index.js'
 import { match } from '../../collate.js'
-import { urlId } from '../../common/url.js'
+import { urlId } from '../../common/util.js'
 
 
 export const RecentProjects = ({
