+++ src/components/project/settings.js
@@ -7,7 +7,7 @@ import { Button } from '../button.js'
 import { ProjectTypeField } from './type-field.js'
 import { notify } from '../../dialog.js'
 import { BASES, TYPES } from '../../common/project.js'
-import { protocolURL } from '../../common/url.js'
+import { protocolURL } from '../../common/util.js'
 import { convert, optimize } from '../../slices/project-files.js'
 import { useDispatch } from 'react-redux'
 import { useIntl } from 'react-intl'
